os : Resolve build warning in os folder make changes in os folder files to resolve the build warning

+++ os/binfmt/elf.c
@@ -247,15 +247,15 @@ static int elf_loadbinary(FAR struct binary_s *binp)
 	}
 
 #ifdef CONFIG_BINFMT_CONSTRUCTORS
-	binp->sections[BIN_CTOR] = loadinfo.ctors;
-	binp->sections[BIN_DTOR] = loadinfo.dtors;
+	binp->sections[BIN_CTOR] = (uint32_t)loadinfo.ctors;
+	binp->sections[BIN_DTOR] = (uint32_t)loadinfo.dtors;
 #endif
 
 #ifdef CONFIG_BINFMT_CONSTRUCTORS
 	/* Save information about constructors and destructors. */
 
 	binp->ctors = loadinfo.ctors;
-	binp->nctors = loadinfo.nctors;
+	binp->nctors = (uint32_t)loadinfo.nctors;
 
 	binp->dtors = loadinfo.dtors;
 	binp->ndtors = loadinfo.ndtors;
